scripts/gemini25/requirements.txt
openai>=1.0.0
anthropic>=0.7.0
google-generativeai>=0.3.0
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.0.0
python-dotenv>=0.19.0
pathlib2>=2.3.0 